requirements.txt
beautifulsoup4==4.12.3
requests==2.32.2
PyYAML==6.0.2